End of day commit. Moved some functions fromt the "screw trap data explore.Rmd" file to their own seperate ".R" files, to use in other scripts.
Getting ready to run data through BTSPAS, although it is currently throwing an error, which I believe is related a problem with the u2copy spline fit, which I believe Thomas has a fix for. Their is this weird u2copy thing in the JAGS code that I can't quite tell what it is doing, but a spline gets fit to it and then it gets log-transformed, so BTSPAS throws an error if the spline fit goes negative. This is happening with the data I tried to run through it. Should be able to work around with Thomas's fix. Need to figure out what that part of the JAGS model is doing. Difficult because I'm not fluent in JAGS.

Started modifying BTSPAS functions based on Thomas Buehrens suggestions (those functions are in .../src/BTSPAS_2).